Fix issue where "--nocolor" is ignored
