+++ src/sqlfluff/cli/commands.py
@@ -7,8 +7,9 @@ import json
 import logging
 import time
 from logging import LogRecord
-from typing import Callable, Tuple, Optional, List, cast
+from typing import Callable, Tuple, Optional, cast
 
+from colorama import Style
 import yaml
 
 import click
@@ -23,16 +24,10 @@ from tqdm import tqdm
 from sqlfluff.cli.autocomplete import shell_completion_enabled, dialect_shell_complete
 
 from sqlfluff.cli.formatters import (
-    format_rules,
-    format_violation,
     format_linting_result_header,
-    format_linting_stats,
-    colorize,
-    format_dialect_warning,
-    format_dialects,
     OutputStreamFormatter,
 )
-from sqlfluff.cli.helpers import cli_table, get_package_version
+from sqlfluff.cli.helpers import get_package_version
 from sqlfluff.cli.outputstream import make_output_stream, OutputStream
 
 # Import from sqlfluff core.
@@ -44,23 +39,24 @@ from sqlfluff.core import (
     SQLFluffUserError,
     dialect_selector,
     dialect_readout,
-    TimingSummary,
 )
 from sqlfluff.core.config import progress_bar_configuration
 
 from sqlfluff.core.enums import FormatType, Color
-from sqlfluff.core.linter import ParsedString
 from sqlfluff.core.plugin.host import get_plugin_manager
-from sqlfluff.core.linter import LintingResult
 
 
 class RedWarningsFilter(logging.Filter):
     """This filter makes all warnings or above red."""
 
+    def __init__(self, formatter: OutputStreamFormatter):
+        super().__init__()
+        self.formatter = formatter
+
     def filter(self, record: logging.LogRecord) -> bool:
         """Filter any warnings (or above) to turn them red."""
         if record.levelno >= logging.WARNING:
-            record.msg = f"{colorize(record.msg, Color.red)} "
+            record.msg = f"{self.formatter.colorize(record.msg, Color.red)} "
         return True
 
 
@@ -83,7 +79,10 @@ class StreamHandlerTqdm(logging.StreamHandler):
 
 
 def set_logging_level(
-    verbosity: int, logger: Optional[logging.Logger] = None, stderr_output: bool = False
+    verbosity: int,
+    formatter: OutputStreamFormatter,
+    logger: Optional[logging.Logger] = None,
+    stderr_output: bool = False,
 ) -> None:
     """Set up logging for the CLI.
 
@@ -111,7 +110,7 @@ def set_logging_level(
     # tamed ANSI colour statements, and return us to normality.
     handler.setFormatter(logging.Formatter("\u001b[0m%(levelname)-10s %(message)s"))
     # Set up a handler to colour warnings red.
-    handler.addFilter(RedWarningsFilter())
+    handler.addFilter(RedWarningsFilter(formatter))
     if logger:
         focus_logger = logging.getLogger(f"sqlfluff.{logger}")
         focus_logger.addHandler(handler)
@@ -292,6 +291,18 @@ def get_config(
     **kwargs,
 ) -> FluffConfig:
     """Get a config object from kwargs."""
+    plain_output = kwargs["nocolor"] or not sys.stdout.isatty()
+
+    def colorize(s: str, color: Optional[Color] = None) -> str:
+        """Optionally use ANSI colour codes to colour a string.
+
+        The name of this function is in American. I'm sorry :(.
+        """
+        if not color or plain_output:
+            return s
+        else:
+            return f"{color.value}{s}{Style.RESET_ALL}"
+
     if kwargs.get("dialect"):
         try:
             # We're just making sure it exists at this stage.
@@ -349,6 +360,7 @@ def get_linter_and_formatter(
         sys.exit(66)
     formatter = OutputStreamFormatter(
         output_stream=output_stream or make_output_stream(cfg),
+        nocolor=cfg.get("nocolor"),
         verbosity=cfg.get("verbose"),
         output_line_length=cfg.get("output_line_length"),
     )
@@ -381,8 +393,8 @@ def version(**kwargs) -> None:
 def rules(**kwargs) -> None:
     """Show the current rules in use."""
     c = get_config(**kwargs, dialect="ansi")
-    lnt, _ = get_linter_and_formatter(c)
-    click.echo(format_rules(lnt), color=c.get("color"))
+    lnt, formatter = get_linter_and_formatter(c)
+    click.echo(formatter.format_rules(lnt), color=c.get("color"))
 
 
 @cli.command()
@@ -390,7 +402,8 @@ def rules(**kwargs) -> None:
 def dialects(**kwargs) -> None:
     """Show the current dialects available."""
     c = get_config(**kwargs, require_dialect=False)
-    click.echo(format_dialects(dialect_readout), color=c.get("color"))
+    _, formatter = get_linter_and_formatter(c)
+    click.echo(formatter.format_dialects(dialect_readout), color=c.get("color"))
 
 
 def dump_file_payload(filename: Optional[str], payload: str):
@@ -505,7 +518,12 @@ def lint(
     formatter.dispatch_config(lnt)
 
     # Set up logging.
-    set_logging_level(verbosity=verbose, logger=logger, stderr_output=non_human_output)
+    set_logging_level(
+        verbosity=verbose,
+        formatter=formatter,
+        logger=logger,
+        stderr_output=non_human_output,
+    )
     # add stdin if specified via lone '-'
     if ("-",) == paths:
         result = lnt.lint_string_wrapped(sys.stdin.read(), fname="stdin")
@@ -522,7 +540,7 @@ def lint(
             )
         except OSError:
             click.echo(
-                colorize(
+                formatter.colorize(
                     f"The path(s) '{paths}' could not be accessed. Check it/they "
                     "exist(s).",
                     Color.red,
@@ -531,7 +549,7 @@ def lint(
             sys.exit(1)
         # Output the final stats
         if verbose >= 1:
-            click.echo(format_linting_stats(result, verbose=verbose))
+            click.echo(formatter.format_linting_stats(result, verbose=verbose))
 
     if format == FormatType.json.value:
         file_output = json.dumps(result.as_records())
@@ -589,42 +607,18 @@ def lint(
     output_stream.close()
     if bench:
         click.echo("==== overall timings ====")
-        click.echo(cli_table([("Clock time", result.total_time)]))
+        click.echo(formatter.cli_table([("Clock time", result.total_time)]))
         timing_summary = result.timing_summary()
         for step in timing_summary:
             click.echo(f"=== {step} ===")
-            click.echo(cli_table(timing_summary[step].items()))
+            click.echo(formatter.cli_table(timing_summary[step].items()))
 
     if not nofail:
         if not non_human_output:
-            _completion_message(config)
+            formatter.completion_message()
         sys.exit(result.stats()["exit code"])
     else:
         sys.exit(0)
-
-
-def _handle_files_with_tmp_or_prs_errors(lint_result: LintingResult) -> int:
-    """Discard lint fixes for files with templating or parse errors.
-
-    Returns 1 if there are any files with templating or parse errors after
-    filtering, else 0. (Intended as a process exit code.)
-    """
-    total_errors, num_filtered_errors = lint_result.count_tmp_prs_errors()
-    lint_result.discard_fixes_for_lint_errors_in_files_with_tmp_or_prs_errors()
-    if total_errors:
-        click.echo(
-            colorize(f"  [{total_errors} templating/parsing errors found]", Color.red)
-        )
-        if num_filtered_errors < total_errors:
-            color = Color.red if num_filtered_errors else Color.green
-            click.echo(
-                colorize(
-                    f"  [{num_filtered_errors} templating/parsing errors "
-                    f'remaining after "ignore"]',
-                    color,
-                )
-            )
-    return 1 if num_filtered_errors else 0
 
 
 def do_fixes(lnt, result, formatter=None, **kwargs):
@@ -724,7 +718,12 @@ def fix(
     formatter.dispatch_config(lnt)
 
     # Set up logging.
-    set_logging_level(verbosity=verbose, logger=logger, stderr_output=fixing_stdin)
+    set_logging_level(
+        verbosity=verbose,
+        formatter=formatter,
+        logger=logger,
+        stderr_output=fixing_stdin,
+    )
 
     # handle stdin case. should output formatted sql to stdout and nothing else.
     if fixing_stdin:
@@ -734,7 +733,7 @@ def fix(
         templater_error = result.num_violations(types=SQLTemplaterError) > 0
         unfixable_error = result.num_violations(types=SQLLintError, fixable=False) > 0
         if not fix_even_unparsable:
-            exit_code = _handle_files_with_tmp_or_prs_errors(result)
+            exit_code = formatter.handle_files_with_tmp_or_prs_errors(result)
 
         if result.num_violations(types=SQLLintError, fixable=True) > 0:
             stdout = result.paths[0].files[0].fix_string()[0]
@@ -743,14 +742,14 @@ def fix(
 
         if templater_error:
             click.echo(
-                colorize(
+                formatter.colorize(
                     "Fix aborted due to unparseable template variables.",
                     Color.red,
                 ),
                 err=True,
             )
             click.echo(
-                colorize(
+                formatter.colorize(
                     "Use --FIX-EVEN-UNPARSABLE' to attempt to fix the SQL anyway.",
                     Color.red,
                 ),
@@ -758,7 +757,10 @@ def fix(
             )
 
         if unfixable_error:
-            click.echo(colorize("Unfixable violations detected.", Color.red), err=True)
+            click.echo(
+                formatter.colorize("Unfixable violations detected.", Color.red),
+                err=True,
+            )
 
         click.echo(stdout, nl=False)
         sys.exit(1 if templater_error or unfixable_error else exit_code)
@@ -774,7 +776,7 @@ def fix(
         )
     except OSError:
         click.echo(
-            colorize(
+            formatter.colorize(
                 f"The path(s) '{paths}' could not be accessed. Check it/they exist(s).",
                 Color.red,
             ),
@@ -783,7 +785,7 @@ def fix(
         sys.exit(1)
 
     if not fix_even_unparsable:
-        exit_code = _handle_files_with_tmp_or_prs_errors(result)
+        exit_code = formatter.handle_files_with_tmp_or_prs_errors(result)
 
     # NB: We filter to linting violations here, because they're
     # the only ones which can be potentially fixed.
@@ -794,7 +796,9 @@ def fix(
             "linting violations found"
         )
         if force:
-            click.echo(f"{colorize('FORCE MODE', Color.red)}: Attempting fixes...")
+            click.echo(
+                f"{formatter.colorize('FORCE MODE', Color.red)}: Attempting fixes..."
+            )
             success = do_fixes(
                 lnt,
                 result,
@@ -822,7 +826,7 @@ def fix(
                 if not success:
                     sys.exit(1)  # pragma: no cover
                 else:
-                    _completion_message(config)
+                    formatter.completion_message()
             elif c == "n":
                 click.echo("Aborting...")
                 exit_code = 1
@@ -832,7 +836,7 @@ def fix(
                 exit_code = 1
     else:
         click.echo("==== no fixable linting violations found ====")
-        _completion_message(config)
+        formatter.completion_message()
 
     error_types = [
         (
@@ -849,20 +853,13 @@ def fix(
 
     if bench:
         click.echo("==== overall timings ====")
-        click.echo(cli_table([("Clock time", result.total_time)]))
+        click.echo(formatter.cli_table([("Clock time", result.total_time)]))
         timing_summary = result.timing_summary()
         for step in timing_summary:
             click.echo(f"=== {step} ===")
-            click.echo(cli_table(timing_summary[step].items()))
+            click.echo(formatter.cli_table(timing_summary[step].items()))
 
     sys.exit(exit_code)
-
-
-def _completion_message(config: FluffConfig) -> None:
-    click.echo(
-        "All Finished"
-        f"{'' if (config.get('nocolor') or not sys.stdout.isatty()) else ' 📜 🎉'}!"
-    )
 
 
 def quoted_presenter(dumper, data):
@@ -965,7 +962,12 @@ def parse(
     formatter.dispatch_config(lnt)
 
     # Set up logging.
-    set_logging_level(verbosity=verbose, logger=logger, stderr_output=non_human_output)
+    set_logging_level(
+        verbosity=verbose,
+        formatter=formatter,
+        logger=logger,
+        stderr_output=non_human_output,
+    )
 
     # TODO: do this better
 
@@ -1001,7 +1003,7 @@ def parse(
 
         # iterative print for human readout
         if format == FormatType.human.value:
-            violations_count = _print_out_violations_and_timing(
+            violations_count = formatter.print_out_violations_and_timing(
                 output_stream, bench, code_only, total_time, verbose, parsed_strings
             )
         else:
@@ -1030,7 +1032,7 @@ def parse(
 
     except OSError:  # pragma: no cover
         click.echo(
-            colorize(
+            formatter.colorize(
                 f"The path '{path}' could not be accessed. Check it exists.",
                 Color.red,
             ),
@@ -1051,52 +1053,6 @@ def parse(
         sys.exit(66)  # pragma: no cover
     else:
         sys.exit(0)
-
-
-def _print_out_violations_and_timing(
-    output_stream: OutputStream,
-    bench: bool,
-    code_only: bool,
-    total_time: float,
-    verbose: int,
-    parsed_strings: List[ParsedString],
-) -> int:
-    """Used by human formatting during the parse."""
-    violations_count = 0
-    timing = TimingSummary()
-
-    for parsed_string in parsed_strings:
-        timing.add(parsed_string.time_dict)
-
-        if parsed_string.tree:
-            output_stream.write(parsed_string.tree.stringify(code_only=code_only))
-        else:
-            # TODO: Make this prettier
-            output_stream.write("...Failed to Parse...")  # pragma: no cover
-
-        violations_count += len(parsed_string.violations)
-        if parsed_string.violations:
-            output_stream.write("==== parsing violations ====")  # pragma: no cover
-        for v in parsed_string.violations:
-            output_stream.write(format_violation(v))  # pragma: no cover
-        if parsed_string.violations:
-            output_stream.write(
-                format_dialect_warning(parsed_string.config.get("dialect"))
-            )
-
-        if verbose >= 2:
-            output_stream.write("==== timings ====")
-            output_stream.write(cli_table(parsed_string.time_dict.items()))
-
-    if verbose >= 2 or bench:
-        output_stream.write("==== overall timings ====")
-        output_stream.write(cli_table([("Clock time", total_time)]))
-        timing_summary = timing.summary()
-        for step in timing_summary:
-            output_stream.write(f"=== {step} ===")
-            output_stream.write(cli_table(timing_summary[step].items()))
-
-    return violations_count
 
 
 # This "__main__" handler allows invoking SQLFluff using "python -m", which
