+++ src/sqlfluff/cli/formatters.py
@@ -1,35 +1,22 @@
 """Defines the formatters for the CLI."""
-
-
 from io import StringIO
-from typing import List, Union
+import sys
+from typing import List, Optional, Tuple, Union
+
+import click
+from colorama import Style
 
 from sqlfluff.cli.helpers import (
-    colorize,
-    cli_table,
     get_package_version,
     get_python_version,
     get_python_implementation,
     pad_line,
+    wrap_field,
 )
 from sqlfluff.cli.outputstream import OutputStream
-from sqlfluff.core import SQLBaseError, FluffConfig, Linter
+from sqlfluff.core import SQLBaseError, FluffConfig, Linter, TimingSummary
 from sqlfluff.core.enums import Color
-from sqlfluff.core.linter import LintedFile
-
-
-def format_filename(
-    filename: str, success: Union[str, bool] = False, success_text: str = "PASS"
-) -> str:
-    """Format filenames."""
-    if isinstance(success, str):
-        status_string = success
-    else:
-        status_string = colorize(
-            success_text if success else "FAIL",
-            Color.green if success else Color.red,
-        )
-    return f"== [{colorize(filename, Color.lightgrey)}] {status_string}"
+from sqlfluff.core.linter import LintedFile, LintingResult, ParsedString
 
 
 def split_string_on_spaces(s: str, line_length: int = 100) -> List[str]:
@@ -54,148 +41,11 @@ def split_string_on_spaces(s: str, line_length: int = 100) -> List[str]:
     return line_buff
 
 
-def format_violation(violation: SQLBaseError, max_line_length: int = 90) -> str:
-    """Format a violation."""
-    if not isinstance(violation, SQLBaseError):  # pragma: no cover
-        raise ValueError(f"Unexpected violation format: {violation}")
-
-    desc: str = violation.desc()
-    line_elem = "   -" if violation.line_no is None else f"{violation.line_no:4d}"
-    pos_elem = "   -" if violation.line_pos is None else f"{violation.line_pos:4d}"
-
-    if violation.ignore:
-        desc = "IGNORE: " + desc  # pragma: no cover
-
-    split_desc = split_string_on_spaces(desc, line_length=max_line_length - 25)
-
-    out_buff = ""
-    # Grey out the violation if we're ignoring it.
-    section_color: Color = Color.lightgrey if violation.ignore else Color.blue
-    for idx, line in enumerate(split_desc):
-        if idx == 0:
-            rule_code = violation.rule_code().rjust(4)
-            if "PRS" in rule_code:
-                section_color = Color.red
-            out_buff += colorize(
-                f"L:{line_elem} | P:{pos_elem} | {rule_code} | ",
-                section_color,
-            )
-        else:
-            out_buff += (
-                "\n"
-                + (" " * 23)
-                + colorize(
-                    "| ",
-                    section_color,
-                )
-            )
-        out_buff += line
-    return out_buff
-
-
-def format_linting_stats(result, verbose=0):
-    """Format a set of stats given a `LintingResult`."""
-    text_buffer = StringIO()
-    all_stats = result.stats()
-    text_buffer.write("==== summary ====\n")
-    if verbose >= 2:
-        output_fields = [
-            "files",
-            "violations",
-            "clean files",
-            "unclean files",
-            "avg per file",
-            "unclean rate",
-            "status",
-        ]
-        special_formats = {"unclean rate": "{0:.0%}"}
-    else:
-        output_fields = ["violations", "status"]
-        special_formats = {}
-    # Generate content tuples, applying special formats for some fields
-    summary_content = [
-        (
-            key,
-            special_formats[key].format(all_stats[key])
-            if key in special_formats
-            else all_stats[key],
-        )
-        for key in output_fields
-    ]
-    # Render it all as a table
-    text_buffer.write(cli_table(summary_content, max_label_width=14))
-    return text_buffer.getvalue()
-
-
 def format_linting_result_header():
     """Format the header of a linting result output."""
     text_buffer = StringIO()
     text_buffer.write("==== readout ====\n")
     return text_buffer.getvalue()
-
-
-def format_config_vals(config_vals):
-    """Format an iterable of config values from a config object."""
-    text_buffer = StringIO()
-    for i, k, v in config_vals:
-        val = "" if v is None else str(v)
-        text_buffer.write(
-            ("    " * i)
-            + colorize(pad_line(str(k) + ":", 20, "left"), color=Color.lightgrey)
-            + pad_line(val, 20, "left")
-            + "\n"
-        )
-    return text_buffer.getvalue()
-
-
-def format_rules(linter: Linter, verbose: int = 0) -> str:
-    """Format the a set of rules given a `Linter`."""
-    text_buffer = StringIO()
-    text_buffer.write("==== sqlfluff - rules ====\n")
-    text_buffer.write(
-        cli_table(
-            linter.rule_tuples(),
-            col_width=80,
-            cols=1,
-            label_color=Color.blue,
-            val_align="left",
-        )
-    )
-    return text_buffer.getvalue()
-
-
-def format_dialects(dialect_readout, verbose=0):
-    """Format the dialects yielded by `dialect_readout`."""
-    text_buffer = StringIO()
-    text_buffer.write("==== sqlfluff - dialects ====\n")
-    readouts = [
-        (
-            dialect.label,
-            f"{dialect.name} dialect [inherits from '{dialect.inherits_from}']",
-        )
-        for dialect in dialect_readout()
-    ]
-    text_buffer.write(
-        cli_table(
-            readouts,
-            col_width=60,
-            cols=1,
-            label_color=Color.blue,
-            val_align="right",
-        )
-    )
-    return text_buffer.getvalue()
-
-
-def format_dialect_warning(dialect):
-    """Output a warning for parsing errors."""
-    return colorize(
-        (
-            "WARNING: Parsing errors found and dialect is set to "
-            f"'{dialect}'. Have you configured your dialect correctly?"
-        ),
-        Color.lightgrey,
-    )
 
 
 class OutputStreamFormatter:
@@ -220,11 +70,13 @@ class OutputStreamFormatter:
     def __init__(
         self,
         output_stream: OutputStream,
+        nocolor: bool,
         verbosity: int = 0,
         filter_empty: bool = True,
         output_line_length: int = 80,
     ):
         self._output_stream = output_stream
+        self.plain_output = nocolor or not sys.stdout.isatty()
         self._verbosity = verbosity
         self._filter_empty = filter_empty
         self.output_line_length = output_line_length
@@ -254,19 +106,19 @@ class OutputStreamFormatter:
                 config_content.append(("dialect", linter.dialect.name))
             config_content += linter.templater.config_pairs()
             text_buffer.write(
-                cli_table(config_content, col_width=30, max_label_width=15)
+                self.cli_table(config_content, col_width=30, max_label_width=15)
             )
             text_buffer.write("\n")
             if linter.config.get("rule_allowlist"):
                 text_buffer.write(
-                    cli_table(
+                    self.cli_table(
                         [("rules", ", ".join(linter.config.get("rule_allowlist")))],
                         col_width=41,
                     )
                 )
             if self._verbosity > 1:
                 text_buffer.write("\n== Raw Config:\n")
-                text_buffer.write(format_config_vals(linter.config.iter_vals()))
+                text_buffer.write(self.format_config_vals(linter.config.iter_vals()))
         return text_buffer.getvalue()
 
     def dispatch_config(self, linter: Linter) -> None:
@@ -277,12 +129,11 @@ class OutputStreamFormatter:
         """Dispatch filenames during a persist operation."""
         # Only show the skip records at higher levels of verbosity
         if self._verbosity >= 2 or result != "SKIP":
-            self._dispatch(format_filename(filename=filename, success=result))
+            self._dispatch(self.format_filename(filename=filename, success=result))
 
-    @staticmethod
-    def _format_path(path: str) -> str:
+    def _format_path(self, path: str) -> str:
         """Format paths."""
-        return f"=== [ path: {colorize(path, Color.lightgrey)} ] ===\n"
+        return f"=== [ path: {self.colorize(path, Color.lightgrey)} ] ===\n"
 
     def dispatch_path(self, path: str) -> None:
         """Dispatch paths for display."""
@@ -294,7 +145,7 @@ class OutputStreamFormatter:
     ) -> None:
         """Dispatch the header displayed before templating."""
         if self._verbosity > 1:
-            self._dispatch(format_filename(filename=fname, success="TEMPLATING"))
+            self._dispatch(self.format_filename(filename=fname, success="TEMPLATING"))
             # This is where we output config diffs if they exist.
             if file_config:
                 # Only output config diffs if there is a config to diff to.
@@ -302,28 +153,30 @@ class OutputStreamFormatter:
                 if config_diff:  # pragma: no cover
                     self._dispatch("   Config Diff:")
                     self._dispatch(
-                        format_config_vals(linter_config.iter_vals(cfg=config_diff))
+                        self.format_config_vals(
+                            linter_config.iter_vals(cfg=config_diff)
+                        )
                     )
 
     def dispatch_parse_header(self, fname: str) -> None:
         """Dispatch the header displayed before parsing."""
         if self._verbosity > 1:
-            self._dispatch(format_filename(filename=fname, success="PARSING"))
+            self._dispatch(self.format_filename(filename=fname, success="PARSING"))
 
     def dispatch_lint_header(self, fname: str) -> None:
         """Dispatch the header displayed before linting."""
         if self._verbosity > 1:
-            self._dispatch(format_filename(filename=fname, success="LINTING"))
+            self._dispatch(self.format_filename(filename=fname, success="LINTING"))
 
     def dispatch_compilation_header(self, templater, message):
         """Dispatch the header displayed before linting."""
         self._dispatch(
-            f"=== [{colorize(templater, Color.lightgrey)}] {message}"
+            f"=== [{self.colorize(templater, Color.lightgrey)}] {message}"
         )  # pragma: no cover
 
     def dispatch_dialect_warning(self, dialect) -> None:
         """Dispatch a warning for dialects."""
-        self._dispatch(format_dialect_warning(dialect))  # pragma: no cover
+        self._dispatch(self.format_dialect_warning(dialect))  # pragma: no cover
 
     def _format_file_violations(
         self, fname: str, violations: List[SQLBaseError]
@@ -335,7 +188,7 @@ class OutputStreamFormatter:
 
         # Only print the filename if it's either a failure or verbosity > 1
         if self._verbosity > 0 or not success:
-            text_buffer.write(format_filename(fname, success=success))
+            text_buffer.write(self.format_filename(fname, success=success))
             text_buffer.write("\n")
 
         # If we have violations, print them
@@ -344,7 +197,9 @@ class OutputStreamFormatter:
             s = sorted(violations, key=lambda v: (v.line_no, v.line_pos))
             for violation in s:
                 text_buffer.write(
-                    format_violation(violation, max_line_length=self.output_line_length)
+                    self.format_violation(
+                        violation, max_line_length=self.output_line_length
+                    )
                 )
                 text_buffer.write("\n")
         str_buffer = text_buffer.getvalue()
@@ -361,3 +216,346 @@ class OutputStreamFormatter:
             fname, linted_file.get_violations(fixable=True if only_fixable else None)
         )
         self._dispatch(s)
+
+    def colorize(self, s: str, color: Optional[Color] = None) -> str:
+        """Optionally use ANSI colour codes to colour a string.
+
+        The name of this function is in American. I'm sorry :(.
+        """
+        if not color or self.plain_output:
+            return s
+        else:
+            return f"{color.value}{s}{Style.RESET_ALL}"
+
+    def cli_table_row(
+        self,
+        fields: List[Tuple[str, str]],
+        col_width,
+        max_label_width=10,
+        sep_char=": ",
+        divider_char=" ",
+        label_color=Color.lightgrey,
+        val_align="right",
+    ) -> str:
+        """Make a row of a CLI table, using wrapped values."""
+        # Do some intel first
+        cols = len(fields)
+        last_col_idx = cols - 1
+        wrapped_fields = [
+            wrap_field(
+                field[0],
+                field[1],
+                width=col_width,
+                max_label_width=max_label_width,
+                sep_char=sep_char,
+            )
+            for field in fields
+        ]
+        max_lines = max(fld["lines"] for fld in wrapped_fields)
+        last_line_idx = max_lines - 1
+        # Make some text
+        buff = StringIO()
+        for line_idx in range(max_lines):
+            for col_idx in range(cols):
+                # Assume we pad labels left and values right
+                fld = wrapped_fields[col_idx]
+                ll = fld["label_list"]
+                vl = fld["val_list"]
+                buff.write(
+                    self.colorize(
+                        pad_line(
+                            ll[line_idx] if line_idx < len(ll) else "",
+                            width=fld["label_width"],
+                        ),
+                        color=label_color,
+                    )
+                )
+                if line_idx == 0:
+                    buff.write(sep_char)
+                else:
+                    buff.write(" " * len(sep_char))
+                buff.write(
+                    pad_line(
+                        vl[line_idx] if line_idx < len(vl) else "",
+                        width=fld["val_width"],
+                        align=val_align,
+                    )
+                )
+                if col_idx != last_col_idx:
+                    buff.write(divider_char)
+                elif line_idx != last_line_idx:
+                    buff.write("\n")
+        return buff.getvalue()
+
+    def cli_table(
+        self,
+        fields,
+        col_width=20,
+        cols=2,
+        divider_char=" ",
+        sep_char=": ",
+        label_color=Color.lightgrey,
+        float_format="{0:.2f}",
+        max_label_width=10,
+        val_align="right",
+    ) -> str:
+        """Make a crude ascii table.
+
+        Assume that `fields` is an iterable of (label, value) pairs.
+        """
+        # First format all the values into strings
+        formatted_fields = []
+        for label, value in fields:
+            label = str(label)
+            if isinstance(value, float):
+                value = float_format.format(value)
+            else:
+                value = str(value)
+            formatted_fields.append((label, value))
+
+        # Set up a buffer to hold the whole table
+        buff = StringIO()
+        while len(formatted_fields) > 0:
+            row_buff: List[Tuple[str, str]] = []
+            while len(row_buff) < cols and len(formatted_fields) > 0:
+                row_buff.append(formatted_fields.pop(0))
+            buff.write(
+                self.cli_table_row(
+                    row_buff,
+                    col_width=col_width,
+                    max_label_width=max_label_width,
+                    sep_char=sep_char,
+                    divider_char=divider_char,
+                    label_color=label_color,
+                    val_align=val_align,
+                )
+            )
+            if len(formatted_fields) > 0:
+                buff.write("\n")
+        return buff.getvalue()
+
+    def format_filename(
+        self,
+        filename: str,
+        success: Union[str, bool] = False,
+        success_text: str = "PASS",
+    ) -> str:
+        """Format filenames."""
+        if isinstance(success, str):
+            status_string = success
+        else:
+            status_string = self.colorize(
+                success_text if success else "FAIL",
+                Color.green if success else Color.red,
+            )
+        return f"== [{self.colorize(filename, Color.lightgrey)}] {status_string}"
+
+    def format_violation(
+        self, violation: SQLBaseError, max_line_length: int = 90
+    ) -> str:
+        """Format a violation."""
+        if not isinstance(violation, SQLBaseError):  # pragma: no cover
+            raise ValueError(f"Unexpected violation format: {violation}")
+
+        desc: str = violation.desc()
+        line_elem = "   -" if violation.line_no is None else f"{violation.line_no:4d}"
+        pos_elem = "   -" if violation.line_pos is None else f"{violation.line_pos:4d}"
+
+        if violation.ignore:
+            desc = "IGNORE: " + desc  # pragma: no cover
+
+        split_desc = split_string_on_spaces(desc, line_length=max_line_length - 25)
+
+        out_buff = ""
+        # Grey out the violation if we're ignoring it.
+        section_color: Color = Color.lightgrey if violation.ignore else Color.blue
+        for idx, line in enumerate(split_desc):
+            if idx == 0:
+                rule_code = violation.rule_code().rjust(4)
+                if "PRS" in rule_code:
+                    section_color = Color.red
+                out_buff += self.colorize(
+                    f"L:{line_elem} | P:{pos_elem} | {rule_code} | ",
+                    section_color,
+                )
+            else:
+                out_buff += (
+                    "\n"
+                    + (" " * 23)
+                    + self.colorize(
+                        "| ",
+                        section_color,
+                    )
+                )
+            out_buff += line
+        return out_buff
+
+    def format_linting_stats(self, result, verbose=0):
+        """Format a set of stats given a `LintingResult`."""
+        text_buffer = StringIO()
+        all_stats = result.stats()
+        text_buffer.write("==== summary ====\n")
+        if verbose >= 2:
+            output_fields = [
+                "files",
+                "violations",
+                "clean files",
+                "unclean files",
+                "avg per file",
+                "unclean rate",
+                "status",
+            ]
+            special_formats = {"unclean rate": "{0:.0%}"}
+        else:
+            output_fields = ["violations", "status"]
+            special_formats = {}
+        # Generate content tuples, applying special formats for some fields
+        summary_content = [
+            (
+                key,
+                special_formats[key].format(all_stats[key])
+                if key in special_formats
+                else all_stats[key],
+            )
+            for key in output_fields
+        ]
+        # Render it all as a table
+        text_buffer.write(self.cli_table(summary_content, max_label_width=14))
+        return text_buffer.getvalue()
+
+    def format_config_vals(self, config_vals):
+        """Format an iterable of config values from a config object."""
+        text_buffer = StringIO()
+        for i, k, v in config_vals:
+            val = "" if v is None else str(v)
+            text_buffer.write(
+                ("    " * i)
+                + self.colorize(
+                    pad_line(str(k) + ":", 20, "left"), color=Color.lightgrey
+                )
+                + pad_line(val, 20, "left")
+                + "\n"
+            )
+        return text_buffer.getvalue()
+
+    def format_rules(self, linter: Linter, verbose: int = 0) -> str:
+        """Format the a set of rules given a `Linter`."""
+        text_buffer = StringIO()
+        text_buffer.write("==== sqlfluff - rules ====\n")
+        text_buffer.write(
+            self.cli_table(
+                linter.rule_tuples(),
+                col_width=80,
+                cols=1,
+                label_color=Color.blue,
+                val_align="left",
+            )
+        )
+        return text_buffer.getvalue()
+
+    def format_dialects(self, dialect_readout, verbose=0):
+        """Format the dialects yielded by `dialect_readout`."""
+        text_buffer = StringIO()
+        text_buffer.write("==== sqlfluff - dialects ====\n")
+        readouts = [
+            (
+                dialect.label,
+                f"{dialect.name} dialect [inherits from '{dialect.inherits_from}']",
+            )
+            for dialect in dialect_readout()
+        ]
+        text_buffer.write(
+            self.cli_table(
+                readouts,
+                col_width=60,
+                cols=1,
+                label_color=Color.blue,
+                val_align="right",
+            )
+        )
+        return text_buffer.getvalue()
+
+    def format_dialect_warning(self, dialect):
+        """Output a warning for parsing errors."""
+        return self.colorize(
+            (
+                "WARNING: Parsing errors found and dialect is set to "
+                f"'{dialect}'. Have you configured your dialect correctly?"
+            ),
+            Color.lightgrey,
+        )
+
+    def handle_files_with_tmp_or_prs_errors(self, lint_result: LintingResult) -> int:
+        """Discard lint fixes for files with templating or parse errors.
+
+        Returns 1 if there are any files with templating or parse errors after
+        filtering, else 0. (Intended as a process exit code.)
+        """
+        total_errors, num_filtered_errors = lint_result.count_tmp_prs_errors()
+        lint_result.discard_fixes_for_lint_errors_in_files_with_tmp_or_prs_errors()
+        if total_errors:
+            click.echo(
+                self.colorize(
+                    f"  [{total_errors} templating/parsing errors found]", Color.red
+                )
+            )
+            if num_filtered_errors < total_errors:
+                color = Color.red if num_filtered_errors else Color.green
+                click.echo(
+                    self.colorize(
+                        f"  [{num_filtered_errors} templating/parsing errors "
+                        f'remaining after "ignore"]',
+                        color,
+                    )
+                )
+        return 1 if num_filtered_errors else 0
+
+    def print_out_violations_and_timing(
+        self,
+        output_stream: OutputStream,
+        bench: bool,
+        code_only: bool,
+        total_time: float,
+        verbose: int,
+        parsed_strings: List[ParsedString],
+    ) -> int:
+        """Used by human formatting during the parse."""
+        violations_count = 0
+        timing = TimingSummary()
+
+        for parsed_string in parsed_strings:
+            timing.add(parsed_string.time_dict)
+
+            if parsed_string.tree:
+                output_stream.write(parsed_string.tree.stringify(code_only=code_only))
+            else:
+                # TODO: Make this prettier
+                output_stream.write("...Failed to Parse...")  # pragma: no cover
+
+            violations_count += len(parsed_string.violations)
+            if parsed_string.violations:
+                output_stream.write("==== parsing violations ====")  # pragma: no cover
+            for v in parsed_string.violations:
+                output_stream.write(self.format_violation(v))  # pragma: no cover
+            if parsed_string.violations:
+                output_stream.write(
+                    self.format_dialect_warning(parsed_string.config.get("dialect"))
+                )
+
+            if verbose >= 2:
+                output_stream.write("==== timings ====")
+                output_stream.write(self.cli_table(parsed_string.time_dict.items()))
+
+        if verbose >= 2 or bench:
+            output_stream.write("==== overall timings ====")
+            output_stream.write(self.cli_table([("Clock time", total_time)]))
+            timing_summary = timing.summary()
+            for step in timing_summary:
+                output_stream.write(f"=== {step} ===")
+                output_stream.write(self.cli_table(timing_summary[step].items()))
+
+        return violations_count
+
+    def completion_message(self) -> None:
+        """Prints message when SQLFluff is finished."""
+        click.echo("All Finished" f"{'' if self.plain_output else ' 📜 🎉'}!")
